chore(axes): add example axis, add docs
